+++ CHANGELOG.md
@@ -11,3 +11,9 @@
 - get profile `client.me`
 - get mailbox `client.messages`
 - get calenders `client.calenders`
+
+## [0.1.1] - (2022-10-21)
+
+### Integrate REST API to get mailbox with pagination
+
+- get mailbox data with next link `client.messages_with_nextlink`